change Them, replace NavBar to another file
