+++ frontend/assets/AppAsset.php
@@ -13,6 +13,7 @@ class AppAsset extends AssetBundle
     public $baseUrl = '@web';
     public $css = [
         'css/site.css',
+        'public/bootstrap.min.css',
     ];
     public $js = [
     ];
